// _src/tsBase/svc/FlowLine.d.ts
import FlowView from './FlowView';
import FlowNode from './FlowNode';
export default class FlowLine {
    private readonly MaxCntCnt1;
    private readonly MinSideCnt2;
    private readonly MinCntCnt3;
    private readonly MinSideSide13;
    private readonly ArrowLen;
    private readonly ArrowWidth;
    private readonly FromTypeAuto;
    private readonly FromTypeV;
    private readonly FromTypeH;
    flowView: FlowView;
    json: Json;
    svg: any;
    fromNode: FlowNode;
    toNode: FlowNode;
    path: any;
    path2: any;
    labelElm: any;
    arrow: any;
    fromType: string;
    isFromTypeV: boolean;
    isFromTypeH: boolean;
    constructor(flowView: FlowView, json: Json, fromNode?: FlowNode, toNode?: FlowNode);
    private _setFromTypeVars;
    private _setEvent;
    setLabel(label: string): void;
    render(): void;
    private _drawLine;
    private _drawArrow;
    getId(): string;
    getFromType(): string;
    setFromType(fromType: string): void;
}